reconstruct att buttons
