+++ components/social-farcaster/proposal/proposal-modals/CreateProposalModal.tsx
@@ -6,7 +6,7 @@ import { formatUnits, TransactionReceipt } from "viem";
 import { PercentPrograssText } from "~/components/common/PercentPrograssText";
 import DialogTabBar from "~/components/layout/tab-view/DialogTabBar";
 import UserWalletSelect from "~/components/portfolio/tokens/UserWalletSelect";
-import { CreateTokenButton } from "~/components/trade/ATTButton";
+import { CreateTokenButton } from "~/components/trade/ATTCreateButton";
 import { Dialog, DialogContent, DialogTrigger } from "~/components/ui/dialog";
 import { Slider } from "~/components/ui/slider";
 import { Text } from "~/components/ui/text";
